fix(decisions): keep proposal-attachment client imports off server-only
The first iteration of the signed-URL upload poisoned two client entry
points with `import 'server-only'`:

- `getStorageObjectByPath` (a new helper) was added to
  `packages/common/src/utils/storage.ts`. Because it imports
  `@op/db/client`, every consumer of the `utils` barrel
  (`./schemaValidator`, `./schemas/instanceData`, etc.) started pulling
  in `server-only` and three Vitest suites began failing at module
  load: `src/client.test.ts`,
  `src/services/decision/schemaValidator.test.ts`,
  `src/services/decision/schemas/instanceData.test.ts`.

- `ProposalAttachments.tsx` (a `'use client'` component) imported the
  new proposal-attachment constants from `@op/common`, which webpack
  traced through `services/index.ts` → `getProposal.ts` →
  `@op/supabase/server` and failed the e2e Next build with
  "You're importing a module that depends on 'next/headers'".

Two surgical changes:

- Move the only db-using helper, `getStorageObjectByPath`, into a new
  `packages/common/src/utils/storageObject.ts`. The two server-side
  callers (`createDocumentResource`, `uploadProposalAttachment`) import
  it directly; the file is intentionally not re-exported from the
  `utils` barrel so client modules can keep importing the barrel
  without dragging `server-only` along.

- Re-export `ALLOWED_PROPOSAL_ATTACHMENT_MIME_TYPES`,
  `MAX_PROPOSAL_ATTACHMENT_FILE_SIZE`, and
  `isAllowedProposalAttachmentMimeType` from
  `packages/common/src/client.ts` (mirroring how the resource
  constants are already exposed), and switch
  `ProposalAttachments.tsx` to import them from `@op/common/client`
  so the same allowlist / size cap stays single-sourced without the
  full `@op/common` barrel landing in the client bundle.

Re-runs of `pnpm w:common test`, `pnpm typecheck`, the
proposal-attachment tRPC routers' Vitest suites, and the e2e Next
build now pass.
