+++ packages/common/src/utils/storage.ts
@@ -1,35 +1,9 @@
-import { db } from '@op/db/client';
-import { objectsInStorage } from '@op/db/schema';
-import { and, eq } from 'drizzle-orm';
-
 // Supabase Storage objects expose `metadata` as opaque jsonb. These helpers
-// safely narrow the well-known fields without an `as` cast.
-
-/**
- * Resolve the storage object row a client just uploaded into via a signed URL.
- * Looked up by (bucket, name) because the client only knows the path it
- * received from sign-upload — not the object UUID.
- */
-export const getStorageObjectByPath = async (input: {
-  bucketId: string;
-  path: string;
-}): Promise<{ id: string; metadata: unknown } | undefined> => {
-  const [row] = await db
-    .select({
-      id: objectsInStorage.id,
-      metadata: objectsInStorage.metadata,
-    })
-    .from(objectsInStorage)
-    .where(
-      and(
-        eq(objectsInStorage.bucketId, input.bucketId),
-        eq(objectsInStorage.name, input.path),
-      ),
-    )
-    .limit(1);
-
-  return row;
-};
+// safely narrow the well-known fields without an `as` cast. They are pure —
+// no database or server-only imports — so the utils barrel stays importable
+// from client-safe entry points. The row lookup that needs the DB lives
+// in `./storageObject.ts`, which is intentionally not re-exported by the
+// utils barrel.
 
 export const getStorageObjectSize = (metadata: unknown): number | null => {
   if (
